requirements.txt
decorator==5.1.0
Django==3.2.9
django-cors-headers==3.10.0
django-debug-toolbar==3.2.2
django-dynamic-db-router==0.3.0
django-filter==21.1
djangorestframework==3.12.4
Flask==1.1.1
Flask-Cors==3.0.8
Flask-MySQLdb==0.2.0
idna==2.8
importlib-metadata==4.8.2
ipython==7.29.0
lazy-object-proxy==1.4.3
Markdown==3.3.6
mysql-connector-python==8.0.26
mysqlclient==2.1.0
requests==2.22.0
SQLAlchemy==1.3.10
sqlparse==0.4.2
virtualenv==16.7.4
